pyproject.toml
[project]
name = "fastai-plain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "pandas>=2.3.2",
    "pandas-stubs>=2.3.0.250703",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.0",
    "timm>=1.0.19",
    "fastai>=2.8.4",
]
